+++ code/ui.datagrid/src/components/DataGrid/index.ts
@@ -1,1 +1,1 @@
-export * from './Grid';
+export * from './DataGrid';
